better integration of cosi in sda
